+++ src/interpreters/Lua_original.rs
@@ -107,7 +107,8 @@ impl Interpreter for Lua_original {
     }
 
     fn execute(&mut self) -> Result<String, SniprunError> {
-        let output = Command::new("lua")
+        let interpreter = Lua_original::get_interpreter_or(&self.data, "lua");
+        let output = Command::new(interpreter)
             .arg(&self.main_file_path)
             .args(&self.get_data().cli_args)
             .output()