+++ src/interpreters/R_original.rs
@@ -96,7 +96,8 @@ impl Interpreter for R_original {
     }
 
     fn execute(&mut self) -> Result<String, SniprunError> {
-        let output = Command::new("Rscript")
+        let interpreter = R_original::get_interpreter_or(&self.data, "Rscript");
+        let output = Command::new(interpreter)
             .arg(&self.main_file_path)
             .args(&self.get_data().cli_args)
             .output()