+++ src/interpreters/TypeScript_original.rs
@@ -112,7 +112,8 @@ impl Interpreter for TypeScript_original {
 
     fn execute(&mut self) -> Result<String, SniprunError> {
         //run th binary and get the std output (or stderr)
-        let output = Command::new("ts-node")
+        let interpreter = TypeScript_original::get_interpreter_or(&self.data, "ts-node");
+        let output = Command::new(interpreter)
             .arg(&self.main_file_path)
             .output()
             .expect("Unable to start process");